♻️ Refactor: 프롬프트 위치 변경
