+++ app/core/chains.py
@@ -2,7 +2,10 @@ from langchain_core.runnables import Runnable
 
 from app.core.llm_client import build_structured_chain, chatOpenAI
 from app.schemas.action_plan_schema import ActionPlanResponse, CandidateResponse, EvaluationResponse, SelectionResponse
-from app.core.prompt_templates.action_plan_templates import get_action_plan_prompt, get_candidate_prompt, get_evaluation_prompt, get_selection_prompt
+from app.core.prompt_templates.candidate_prompt import get_candidate_prompt
+from app.core.prompt_templates.evaluation_prompt import get_evaluation_prompt
+from app.core.prompt_templates.final_select_prompt import get_selection_prompt
+from app.core.prompt_templates.action_plan_prompt import get_action_plan_prompt
 
 candidate_chain: Runnable = build_structured_chain(
     chatOpenAI,
@@ -28,4 +31,3 @@ action_plan_chain: Runnable = build_structured_chain(
     ActionPlanResponse,
 )
 
-swot_chain: Runnable
