+++ app/core/prompt_templates/action_plan_prompt.py
@@ -1,0 +1,155 @@
+import json
+from langchain_core.prompts import ChatPromptTemplate
+
+# 4. To-Do List 생성 Chain
+def get_action_plan_prompt() -> ChatPromptTemplate:
+    # 1단계: SWOT 데이터 (베이커리 예시 유지)
+    example_swot = {
+        "strengths": [{"id": "S1", "keyword": "당일 생산 수제 빵의 품질"}],
+        "weaknesses": [{"id": "W1", "keyword": "온라인 홍보 부족"}],
+        "opportunities": [{"id": "O1", "keyword": "인근 지하철역 출근길 유동인구 풍부"}],
+        "threats": [{"id": "T1", "keyword": "대형 프랜차이즈 빵집 입점 예정"}]
+    }
+
+    # 3단계에서 최종 선정된 3가지 전략 예시
+    example_selected = [
+        {
+            "id": 1,
+            "title": "출근길 타겟 '수제 모닝 샌드위치' 출시",
+            "tags": ["#매출증대", "#난이도하", "#상품개발"],
+            "related_swot": ["S1", "O1"],
+            "final_reason": "지하철역 인근의 풍부한 유동인구라는 지리적 이점을 충분히 활용하기 위해서는 출근 시간대 바쁜 직장인들을 겨냥한 특화 메뉴 도입이 시급합니다. \n매장의 핵심 경쟁력인 당일 생산 수제 빵의 품질을 샌드위치라는 간편식 형태로 전환하면, 아침 시간대 고객의 구매 허들을 낮추면서도 높은 만족도를 제공할 수 있습니다. \n이는 단순한 신규 고객 유입을 넘어 오전 시간대의 안정적인 매출 캐시카우를 확보함으로써 전체적인 수익 구조를 탄탄하게 만드는 전략적 전환점이 될 것입니다."
+        
+        },
+        {
+            "id": 2,
+            "title": "인스타그램 '빵 나오는 시간' 인증 이벤트",
+            "tags": ["#인지도제고", "#난이도하", "#마케팅"],
+            "related_swot": ["W1", "O1"],
+            "final_reason": "풍부한 유동인구에도 불구하고 온라인 노출이 부족하여 잠재 고객의 발길이 끊기는 현상을 해결하기 위해, SNS 참여형 이벤트를 통한 디지털 가시성 확보가 반드시 필요합니다. \n인스타그램을 활발히 이용하는 젊은 층을 타겟으로 빵 나오는 시간을 인증하게 유도함으로써, 별도의 마케팅 비용 없이도 자발적인 홍보 데이터를 축적하고 브랜드 인지도를 빠르게 확산시킬 수 있습니다. \n단기적인 방문 증대 효과는 물론, 온라인상의 신뢰도를 구축하여 장기적으로 매장의 경쟁력을 비약적으로 높여주는 효율적인 솔루션입니다."
+        },
+        {
+            "id": 3,
+            "title": "프랜차이즈 대비 '프리미엄 수제 빵' 브랜딩",
+            "tags": ["#신규고객", "#난이도중", "#고객관리"],
+            "related_swot": ["S1", "T1"],
+            "final_reason": "인근 대형 프랜차이즈의 입점이라는 위협 요소에 대응하기 위해, 가격 경쟁이 아닌 제품의 본질적 가치를 강조하는 프리미엄 브랜딩 전략이 핵심입니다. \n공장형 제품으로는 따라올 수 없는 당일 생산 수제 빵의 정체성을 전면에 내세움으로써 가치 소비를 중시하는 단골 고객층을 확실히 선점해야 합니다. \n브랜딩을 위한 초기 기획 과정이 수반되지만, 이는 프랜차이즈의 공세에도 흔들리지 않는 독보적인 팬층을 확보하고 장기적인 생존력을 강화하는 필수적인 투자 단계가 될 것입니다."
+        }
+    ]
+
+    # 최종 출력될 실행 계획 예시
+    example_output = [
+        {
+            "id": 1,
+            "title": "출근길 타겟 '수제 모닝 샌드위치' 출시",
+            "action_plan": [
+                {
+                    "step": 1,
+                    "title": "핵심 메뉴 구성 및 패키징 확정",
+                    "description": "기존 수제 식빵을 활용한 3분 내 서빙 가능한 레시피를 확정하고, 이동 중 취식이 편한 종이 패키지를 선정합니다.",
+                    "expected_outcome": "조리 효율성이 확보된 표준 레시피와 고객 편의성을 고려한 패키징 시안이 준비됩니다."
+                },
+                {
+                    "step": 2,
+                    "title": "얼리버드 타임 세일 프로모션 세팅",
+                    "description": "오전 7시~9시 사이 샌드위치 구매 시 아메리카노 할인 등 출근족을 위한 가격 혜택을 설계합니다.",
+                    "expected_outcome": "경쟁사 대비 아침 시간대 방문 유인책을 확보하여 신규 유입을 유도합니다."
+                },
+                {
+                    "step": 3,
+                    "title": "매장 외부 가시성 강화(X-배너 설치)",
+                    "description": "지하철역 출구 방향 유동인구가 볼 수 있도록 매장 입구에 '갓 구운 모닝 샌드위치' 홍보 배너를 설치합니다.",
+                    "expected_outcome": "매장 앞 유입률과 아침 시간대 신규 고객 수가 즉각적으로 증가합니다."
+                }
+            ]
+        },
+        {
+            "id": 2,
+            "title": "인스타그램 '빵 나오는 시간' 인증 이벤트",
+            "action_plan": [
+                {
+                    "step": 1,
+                    "title": "이벤트 홍보물 및 참여 가이드 제작",
+                    "description": "매장 내 '빵 나오는 시간표'와 함께 SNS 인증 시 혜택(예: 쿠키 증정)을 알리는 안내문을 비치합니다.",
+                    "expected_outcome": "내방 고객들이 이벤트 내용을 인지하고 참여할 수 있는 환경이 조성됩니다."
+                },
+                {
+                    "step": 2,
+                    "title": "공식 계정 활성화 및 해시태그 설정",
+                    "description": "지역 맛집 관련 해시태그를 설정하고, 고객이 올린 인증샷을 공식 계정에서 리포스팅하며 소통합니다.",
+                    "expected_outcome": "지역 기반 검색 결과에 매장 노출 빈도가 높아지며 온라인 바이럴이 시작됩니다."
+                },
+                {
+                    "step": 3,
+                    "title": "베스트 리뷰어 선정 및 재방문권 증정",
+                    "description": "매주 정성스러운 후기를 남긴 고객 3명을 선정하여 재방문 시 사용 가능한 할인권을 DM으로 발송합니다.",
+                    "expected_outcome": "고품질 후기 데이터가 축적되며 고객의 자발적인 재방문율이 상승합니다."
+                }
+            ]
+        },
+        {
+            "id": 3,
+            "title": "프랜차이즈 대비 '프리미엄 수제 빵' 브랜딩",
+            "action_plan": [
+                {
+                    "step": 1,
+                    "title": "장인 정신 소구 문구 및 스토리 개발",
+                    "description": "'24시간 저온 숙성', '천연 효모 사용' 등 제품의 차별화된 가치를 담은 브랜드 카피를 개발합니다.",
+                    "expected_outcome": "프랜차이즈와 다른 우리 매장만의 고유한 브랜드 메시지가 정립됩니다."
+                },
+                {
+                    "step": 2,
+                    "title": "프리미엄 패키징 및 진열 구획 개선",
+                    "description": "시그니처 제품에 고급스러운 라벨지를 부착하고, 조명과 진열대를 조정하여 제품이 돋보이게 연출합니다.",
+                    "expected_outcome": "고객이 체감하는 시각적 품질이 향상되어 가치 소비를 지향하는 고객을 만족시킵니다."
+                },
+                {
+                    "step": 3,
+                    "title": "멤버십 포인트 및 단골 우대 혜택 강화",
+                    "description": "단골 고객을 위한 선결제 시스템이나 신메뉴 우선 시식 기회를 제공하는 VIP 등급제를 도입합니다.",
+                    "expected_outcome": "프랜차이즈 침투에도 이탈하지 않는 강력한 로열티를 가진 단골 팬층이 형성됩니다."
+                }
+            ]
+        }
+    ]
+
+    # AI 프롬프트
+    return ChatPromptTemplate.from_messages([
+        ("system", """
+        너는 선정된 비즈니스 전략을 실질적인 운영 프로세스로 전환하는 '현장 중심형 실행 컨설턴트'다. 
+        사장님이 복잡한 고민 없이 즉시 실행할 수 있도록 3단계의 구체적인 액션 아이템을 도출하라.
+
+        ### 목표
+        최종 선정된 각 전략에 대해 **[준비 - 실행 - 고도화]**의 흐름을 가진 3단계 실행 계획(Action Plan)을 수립하라.
+
+        ### 작성 규칙
+        1. **현실 가능성**: 소상공인의 한정된 인력과 예산으로도 일주일 내에 착수할 수 있는 현실적인 방안이어야 한다.
+        2. **구체성**: '홍보를 잘 하자' 같은 추상적인 문구는 배제하고, '어떤 매체를 통해', '어떤 문구로', '어떻게' 할지 명확히 서술하라.
+        3. **단계별 흐름**:
+            - 1단계(준비): 상품 기획, 재료 준비, 인프라 세팅 등 기초 단계
+            - 2단계(실행): 실제 출시, 프로모션 진행, 고객 응대 등 현장 적용 단계
+            - 3단계(확산/관리): 후기 유도, 단골 관리, 데이터 분석 등 지속 가능 단계
+        4. **가시적 결과 명시**: 각 단계(`step`)마다 해당 행동을 완료했을 때 사장님이 얻을 수 있는 '기대 결과(`expected_outcome`)'를 한 문장으로 구체적으로 포함하라.
+        5. **어조**: 신뢰감을 주는 비즈니스 어투를 사용하되, 실천을 독려하는 긍정적인 뉘앙스를 유지하라.
+        """),
+
+        # Few-shot 데이터
+        ("human", f"""
+        ### 분석 데이터:
+        - SWOT 데이터: {json.dumps(example_swot, ensure_ascii=False)}
+        - 최종 선정 전략: {json.dumps(example_selected, ensure_ascii=False)}
+
+        위 전략들에 대해 사장님이 바로 행동에 옮길 수 있는 3단계 실행 계획을 세밀하게 작성해줘.
+        """),
+        
+        ("ai", json.dumps(example_output, ensure_ascii=False)),
+
+        # 실제 요청
+        ("human", """
+        ### 분석 데이터:
+        - SWOT 데이터: {swot_json}
+        - 최종 선정 전략: {final_selected_strategies}
+
+        위 전략들에 대해 사장님이 바로 행동에 옮길 수 있는 3단계 실행 계획을 세밀하게 작성해줘.
+        """),
+    ])
