--- app/core/prompt_templates/action_plan_templates.py
@@ -1,483 +1,0 @@
-import json
-from langchain_core.prompts import ChatPromptTemplate
-
-# 1. 후보 전략 선정 Chain
-def get_candidate_prompt() -> ChatPromptTemplate:
-    # Few-shot Input (베이커리 예시)
-    example_input = {
-        "strengths": [{"id": "S1", "keyword": "당일 생산 수제 빵의 품질"}],
-        "weaknesses": [{"id": "W1", "keyword": "온라인 홍보 부족"}],
-        "opportunities": [{"id": "O1", "keyword": "인근 지하철역 출근길 유동인구 풍부"}],
-        "threats": [{"id": "T1", "keyword": "대형 프랜차이즈 빵집 입점 예정"}]
-    }
-    # Few-shot Output
-    example_output = [
-        {
-            "id": 1,
-            "title": "출근길 타겟 '수제 모닝 샌드위치' 출시",
-            "tags": ["#매출증대", "#난이도하", "#상품개발"],
-            "related_swot": ["S1", "O1"],
-            "reason": "수제 빵의 품질(S1)과 풍부한 출근길 유동인구(O1)를 결합하여 아침 시간대 구매를 유도하는 SO 전략이다."
-        },
-        {
-            "id": 2,
-            "title": "인스타그램 '빵 나오는 시간' 인증 이벤트",
-            "tags": ["#인지도제고", "#난이도하", "#마케팅"],
-            "related_swot": ["W1", "O1"],
-            "reason": "홍보 부족(W1)을 해결하기 위해 유동인구(O1)를 대상으로 SNS 참여를 유도하여 온라인 노출을 극대화하는 WO 전략이다."
-        },
-        {
-            "id": 3,
-            "title": "프랜차이즈 대비 '프리미엄 수제 빵' 브랜딩",
-            "tags": ["#신규고객", "#난이도중", "#고객관리"],
-            "related_swot": ["S1", "T1"],
-            "reason": "대형 프랜차이즈(T1)에 맞서 수제 빵의 품질(S1)을 강조하여 장인 정신이 담긴 매장임을 차별화하는 ST 전략이다."
-        },
-        {
-            "id": 4,
-            "title": "인근 지역 커뮤니티 타겟 마케팅 강화",
-            "tags": ["#바이럴", "#난이도중", "#마케팅"],
-            "related_swot": ["W1", "T1"],
-            "reason": "부족한 홍보(W1)를 보완하고 경쟁사(T1)의 유입 독점을 막기 위해 지역 맘카페 등에 신뢰 기반 마케팅을 펼치는 WT 전략이다."
-        },
-        {
-            "id": 5,
-            "title": "출근 시간대 '사전 예약 및 픽업' 서비스",
-            "tags": ["#운영효율", "#난이도중", "#시설개선"],
-            "related_swot": ["S1", "O1"],
-            "reason": "빵의 품질(S1)을 아는 바쁜 직장인(O1)을 위해 대기 시간을 줄여 편의성을 제공하는 고도화된 SO 전략이다."
-        }
-    ]
-    
-    # AI 프롬프트
-    return ChatPromptTemplate.from_messages([
-        ("system", """
-        너는 주어진 SWOT 분석 결과를 바탕으로, 소상공인을 위한 구체적이고 실행 가능한 비즈니스 전략을 제안하는 최고의 데이터 기반 전략 컨설턴트다.
-
-        ### 목표
-        입력된 SWOT 분석(JSON)을 바탕으로, 사용자가 실행할 수 있는 후보 솔루션을 3~5개 생성하라. 
-        이 단계에서는 최종 선정 전, 다양한 전략적 조합(SO, WO, ST, WT)을 통해 아이디어를 발산하는 것에 집중한다.
-
-        ### 핵심 규칙
-        1. **전략적 사고**: SO(강점-기회), WO(약점-기회), ST(강점-위협), WT(약점-위협) 등의 조합을 활용하여 솔루션을 도출하라.
-        2. **구체성**: `title`은 사용자가 즉시 이해할 수 있도록 구체적인 액션을 담아 30자 이내로 작성하라.
-        3. **구조화된 태그**: `tags`는 '#목표', '#난이도', '#카테고리' 순서로 반드시 3개를 포함하라.
-            - #목표 예시: #매출증대, #신규고객, #객단가UP, #인지도제고
-            - #난이도 예시: #난이도하, #난이도중, #난이도상
-            - #카테고리 예시: #마케팅, #운영, #고객관리, #상품개발, #시설개선
-        4. **근거 제시**: `related_swot`에는 솔루션 도출에 가장 큰 영향을 준 SWOT 항목의 `id`를 배열로 포함하라.
-        5. **생성 이유 명시**: `reason` 필드에 해당 SWOT 항목들을 왜 그렇게 조합했는지, 그리고 이 전략이 어떤 기대 효과를 갖는지 2문장 내외로 친절하게 설명하라.
-        6. **ID 생성**: Id는 1부터 시작하여 생성되는 후보군의 순서에 따라 **1씩 증가하는 정수(Integer)**로 부여하라. (예: 1, 2, 3...)
-        7. **후보군 개수**: 반드시 3개 이상, 5개 이하의 솔루션을 생성하라.
-        """),
-
-        # Few-shot: Human 예시
-        ("human", f"### 입력된 SWOT 분석 데이터:\n{json.dumps(example_input, ensure_ascii=False)}"),
-        
-        # Few-shot: AI 응답 예시
-        ("ai", json.dumps(example_output, ensure_ascii=False)),
-
-        # 실제 요청
-        ("human", """
-        ### 입력된 SWOT 분석 데이터:
-        {swot_json}        
-
-        이제 위 형식에 맞춰 후보군을 생성해줘.
-        """),
-    ])
-
-# 2. 후보 전략 평가 Chain
-def get_evaluation_prompt() -> ChatPromptTemplate:
-    # Few-shot Input
-    example_swot = {
-        "strengths": [{"id": "S1", "keyword": "당일 생산 수제 빵의 품질"}],
-        "weaknesses": [{"id": "W1", "keyword": "온라인 홍보 부족"}],
-        "opportunities": [{"id": "O1", "keyword": "인근 지하철역 출근길 유동인구 풍부"}],
-        "threats": [{"id": "T1", "keyword": "대형 프랜차이즈 빵집 입점 예정"}]
-    }
-    
-    
-    example_candidates = [
-        {
-            "id": 1,
-            "title": "출근길 타겟 '수제 모닝 샌드위치' 출시",
-            "tags": ["#매출증대", "#난이도하", "#상품개발"],
-            "related_swot": ["S1", "O1"],
-            "reason": "수제 빵의 품질(S1)과 풍부한 출근길 유동인구(O1)를 결합하여 아침 시간대 구매를 유도하는 SO 전략이다."
-        },
-        {
-            "id": 2,
-            "title": "인스타그램 '빵 나오는 시간' 인증 이벤트",
-            "tags": ["#인지도제고", "#난이도하", "#마케팅"],
-            "related_swot": ["W1", "O1"],
-            "reason": "홍보 부족(W1)을 해결하기 위해 유동인구(O1)를 대상으로 SNS 참여를 유도하여 온라인 노출을 극대화하는 WO 전략이다."
-        },
-        {
-            "id": 3,
-            "title": "프랜차이즈 대비 '프리미엄 수제 빵' 브랜딩",
-            "tags": ["#신규고객", "#난이도중", "#고객관리"],
-            "related_swot": ["S1", "T1"],
-            "reason": "대형 프랜차이즈(T1)에 맞서 수제 빵의 품질(S1)을 강조하여 장인 정신이 담긴 매장임을 차별화하는 ST 전략이다."
-        },
-        {
-            "id": 4,
-            "title": "인근 지역 커뮤니티 타겟 마케팅 강화",
-            "tags": ["#바이럴", "#난이도중", "#마케팅"],
-            "related_swot": ["W1", "T1"],
-            "reason": "부족한 홍보(W1)를 보완하고 경쟁사(T1)의 유입 독점을 막기 위해 지역 맘카페 등에 신뢰 기반 마케팅을 펼치는 WT 전략이다."
-        },
-        {
-            "id": 5,
-            "title": "출근 시간대 '사전 예약 및 픽업' 서비스",
-            "tags": ["#운영효율", "#난이도중", "#시설개선"],
-            "related_swot": ["S1", "O1"],
-            "reason": "빵의 품질(S1)을 아는 바쁜 직장인(O1)을 위해 대기 시간을 줄여 편의성을 제공하는 고도화된 SO 전략이다."
-        }
-    ]
-    
-    # Few-shot Output
-    example_output = [
-        {"id": 1, "impactScore": 9, "effortScore": 2, "evaluation": "이미 확보된 품질과 유동인구를 즉시 매출로 연결하며, 메뉴 구성 외 추가 비용이 적어 효율적이다."},
-        {"id": 2, "impactScore": 7, "effortScore": 1, "evaluation": "마케팅 비용 없이 고객 참여만으로 온라인 인지도를 빠르게 쌓을 수 있어 실행이 매우 쉽다."},
-        {"id": 3, "impactScore": 8, "effortScore": 4, "evaluation": "경쟁사와 차별화되는 장기적 자산이 되지만, 브랜딩을 위한 스토리텔링과 디자인 작업이 수반되어야 한다."},
-        {"id": 4, "impactScore": 6, "effortScore": 3, "evaluation": "지역 밀착형 홍보로 확실한 단골을 잡을 수 있으나, 커뮤니티 관리와 신뢰 형성에 시간이 소요된다."},
-        {"id": 5, "impactScore": 5, "effortScore": 6, "evaluation": "편의성은 증대되나 예약 관리 시스템 도입과 인력 동선 조정 등 운영상 복잡성이 발생할 수 있다."}
-    ]
-
-    # AI 프롬프트
-    return ChatPromptTemplate.from_messages([
-        ("system", """
-        너는 생성된 비즈니스 전략 후보군을 객관적인 데이터와 시장 상황을 바탕으로 평가하는 전략 분석가다.
-
-        ### 목표
-        제시된 3~5개의 전략 후보군에 대해 '기대 효과(Impact)'와 '실행 난이도(Effort)'를 평가하라.
-
-        ### 평가 기준
-        1. **기대 효과 (Impact)**: 
-            - 0~10점 (높을수록 좋음)
-            - 매출 증대 가능성, 신규 고객 유입량, 브랜드 인지도 향상 등을 고려하라.
-        2. **실행 난이도 (Effort)**: 
-            - 0~10점 (낮을수록 실행이 쉬움)
-            - 소요 시간, 자금 투자 규모, 기술적 복잡성, 인력 필요도를 고려하라.
-            - 점수가 낮을수록 'Low Effort(쉬움)'이며, 점수가 높을수록 'High Effort(어려움)'이다.
-
-        ### 핵심 규칙
-        - `evaluation`은 왜 해당 점수를 부여했는지 전략적 근거를 2문장 내외로 서술하라.
-        - SWOT 분석 결과와 전략의 연관성을 고려하여 평가의 객관성을 유지하라.
-        """),
-
-        # Few-shot: Human 예시 (SWOT + 후보군)
-        ("human", f"""
-        ### SWOT 데이터:
-        {json.dumps(example_swot, ensure_ascii=False)}
-
-        ### 전략 후보군:
-        {json.dumps(example_candidates, ensure_ascii=False)}
-        """),
-        
-        # Few-shot: AI 응답 예시 (평가 결과)
-        ("ai", json.dumps(example_output, ensure_ascii=False)),
-
-        # 실제 사용자 요청
-        ("human", """
-        ### SWOT 데이터:
-        {swot_json}
-
-        ### 전략 후보군:
-        {candidate_list}
-
-        위 후보군들에 대해 'Impact'와 'Effort' 점수를 산출하고 그 이유를 분석해줘.
-        """),
-    ])
-
-# 3. 최종 선정 Chain
-def get_selection_prompt() -> ChatPromptTemplate:
-    # 1단계: SWOT 데이터 정의 (베이커리 예시)
-    example_swot = {
-        "strengths": [{"id": "S1", "keyword": "당일 생산 수제 빵의 품질"}],
-        "weaknesses": [{"id": "W1", "keyword": "온라인 홍보 부족"}],
-        "opportunities": [{"id": "O1", "keyword": "인근 지하철역 출근길 유동인구 풍부"}],
-        "threats": [{"id": "T1", "keyword": "대형 프랜차이즈 빵집 입점 예정"}]
-    }
-
-    # 2단계: 1번 프롬프트에서 생성된 후보군 (5개)
-    example_candidates = [
-        {
-            "id": 1,
-            "title": "출근길 타겟 '수제 모닝 샌드위치' 출시",
-            "tags": ["#매출증대", "#난이도하", "#상품개발"],
-            "related_swot": ["S1", "O1"],
-            "reason": "수제 빵의 품질(S1)과 풍부한 출근길 유동인구(O1)를 결합하여 아침 시간대 구매를 유도하는 SO 전략이다."
-        },
-        {
-            "id": 2,
-            "title": "인스타그램 '빵 나오는 시간' 인증 이벤트",
-            "tags": ["#인지도제고", "#난이도하", "#마케팅"],
-            "related_swot": ["W1", "O1"],
-            "reason": "홍보 부족(W1)을 해결하기 위해 유동인구(O1)를 대상으로 SNS 참여를 유도하여 온라인 노출을 극대화하는 WO 전략이다."
-        },
-        {
-            "id": 3,
-            "title": "프랜차이즈 대비 '프리미엄 수제 빵' 브랜딩",
-            "tags": ["#신규고객", "#난이도중", "#고객관리"],
-            "related_swot": ["S1", "T1"],
-            "reason": "대형 프랜차이즈(T1)에 맞서 수제 빵의 품질(S1)을 강조하여 장인 정신이 담긴 매장임을 차별화하는 ST 전략이다."
-        },
-        {
-            "id": 4,
-            "title": "인근 지역 커뮤니티 타겟 마케팅 강화",
-            "tags": ["#바이럴", "#난이도중", "#마케팅"],
-            "related_swot": ["W1", "T1"],
-            "reason": "부족한 홍보(W1)를 보완하고 경쟁사(T1)의 유입 독점을 막기 위해 지역 맘카페 등에 신뢰 기반 마케팅을 펼치는 WT 전략이다."
-        },
-        {
-            "id": 5,
-            "title": "출근 시간대 '사전 예약 및 픽업' 서비스",
-            "tags": ["#운영효율", "#난이도중", "#시설개선"],
-            "related_swot": ["S1", "O1"],
-            "reason": "빵의 품질(S1)을 아는 바쁜 직장인(O1)을 위해 대기 시간을 줄여 편의성을 제공하는 고도화된 SO 전략이다."
-        }
-    ]
-
-    # 3단계: 2번 프롬프트에서 생성된 평가 결과 (5개)
-    example_evaluation = [
-        {"id": 1, "impactScore": 9, "effortScore": 2, "evaluation": "이미 확보된 품질과 유동인구를 즉시 매출로 연결하며, 메뉴 구성 외 추가 비용이 적어 효율적이다."},
-        {"id": 2, "impactScore": 7, "effortScore": 1, "evaluation": "마케팅 비용 없이 고객 참여만으로 온라인 인지도를 빠르게 쌓을 수 있어 실행이 매우 쉽다."},
-        {"id": 3, "impactScore": 8, "effortScore": 4, "evaluation": "경쟁사와 차별화되는 장기적 자산이 되지만, 브랜딩을 위한 스토리텔링과 디자인 작업이 수반되어야 한다."},
-        {"id": 4, "impactScore": 6, "effortScore": 3, "evaluation": "지역 밀착형 홍보로 확실한 단골을 잡을 수 있으나, 커뮤니티 관리와 신뢰 형성에 시간이 소요된다."},
-        {"id": 5, "impactScore": 5, "effortScore": 6, "evaluation": "편의성은 증대되나 예약 관리 시스템 도입과 인력 동선 조정 등 운영상 복잡성이 발생할 수 있다."}
-    ]
-
-    # 최종 결과물: 가성비(Impact 높고 Effort 낮은) 및 다양성 고려한 상위 3개 선정
-    example_output = [
-        {
-            "id": 1,
-            "title": "출근길 타겟 '수제 모닝 샌드위치' 출시",
-            "tags": ["#매출증대", "#난이도하", "#상품개발"],
-            "related_swot": ["S1", "O1"],
-            "final_reason": "지하철역 인근의 풍부한 유동인구라는 지리적 이점을 충분히 활용하기 위해서는 출근 시간대 바쁜 직장인들을 겨냥한 특화 메뉴 도입이 시급합니다. \n매장의 핵심 경쟁력인 당일 생산 수제 빵의 품질을 샌드위치라는 간편식 형태로 전환하면, 아침 시간대 고객의 구매 허들을 낮추면서도 높은 만족도를 제공할 수 있습니다. \n이는 단순한 신규 고객 유입을 넘어 오전 시간대의 안정적인 매출 캐시카우를 확보함으로써 전체적인 수익 구조를 탄탄하게 만드는 전략적 전환점이 될 것입니다."
-        
-        },
-        {
-            "id": 2,
-            "title": "인스타그램 '빵 나오는 시간' 인증 이벤트",
-            "tags": ["#인지도제고", "#난이도하", "#마케팅"],
-            "related_swot": ["W1", "O1"],
-            "final_reason": "풍부한 유동인구에도 불구하고 온라인 노출이 부족하여 잠재 고객의 발길이 끊기는 현상을 해결하기 위해, SNS 참여형 이벤트를 통한 디지털 가시성 확보가 반드시 필요합니다. \n인스타그램을 활발히 이용하는 젊은 층을 타겟으로 빵 나오는 시간을 인증하게 유도함으로써, 별도의 마케팅 비용 없이도 자발적인 홍보 데이터를 축적하고 브랜드 인지도를 빠르게 확산시킬 수 있습니다. \n단기적인 방문 증대 효과는 물론, 온라인상의 신뢰도를 구축하여 장기적으로 매장의 경쟁력을 비약적으로 높여주는 효율적인 솔루션입니다."
-        },
-        {
-            "id": 3,
-            "title": "프랜차이즈 대비 '프리미엄 수제 빵' 브랜딩",
-            "tags": ["#신규고객", "#난이도중", "#고객관리"],
-            "related_swot": ["S1", "T1"],
-            "final_reason": "인근 대형 프랜차이즈의 입점이라는 위협 요소에 대응하기 위해, 가격 경쟁이 아닌 제품의 본질적 가치를 강조하는 프리미엄 브랜딩 전략이 핵심입니다. \n공장형 제품으로는 따라올 수 없는 당일 생산 수제 빵의 정체성을 전면에 내세움으로써 가치 소비를 중시하는 단골 고객층을 확실히 선점해야 합니다. \n브랜딩을 위한 초기 기획 과정이 수반되지만, 이는 프랜차이즈의 공세에도 흔들리지 않는 독보적인 팬층을 확보하고 장기적인 생존력을 강화하는 필수적인 투자 단계가 될 것입니다."
-        }
-    ]
-
-    
-    # AI 프롬프트
-    return ChatPromptTemplate.from_messages([
-        ("system", """
-        너는 여러 비즈니스 전략 중 소상공인에게 가장 큰 이익을 줄 수 있는 핵심 솔루션을 최대 3개까지 선정하는 수석 전략 컨설턴트다.
-
-        ### 목표
-        평가된 후보군 중 가성비(High Impact, Low Effort)가 뛰어난 **핵심 솔루션(Core Solutions)**을 우선순위에 따라 **최대 3개** 선정하여 리스트 형식으로 반환하라.
-
-        ### 선정 기준
-        1. **제 1원칙 (Impact)**: 기대 효과(Impact Score)가 가장 높은 솔루션을 우선한다.
-        2. **제 2원칙 (Effort)**: 기대 효과 점수가 동일하다면, 실행 난이도(Effort Score)가 가장 낮은 솔루션을 선택한다.
-        3. **제 3원칙 (Diversity)**: 가능한 한 서로 다른 SWOT 요소(S, W, O, T)를 해결하는 솔루션들을 조합하여 다각적인 전략을 구성하라.
-        3. **최종 검증**: 선정된 솔루션이 주어진 SWOT 데이터와 논리적으로 완벽하게 연결되는지 확인하라.
-         
-        ### final_reason 작성 지침 (분석적 접근)
-        사장님이 전략의 타당성을 이성적으로 납득할 수 있도록 다음 규칙을 엄수하라.
-        1. **전문적이고 정중한 톤**: 예의는 갖추되, 감성적인 위로보다는 전문가적인 통찰을 전달하라.
-        2. **논리적 인과관계**: '강점 활용 - 기회 포착 - 약점 보완' 등 SWOT 분석 결과가 어떻게 수익으로 연결되는지 4~5줄로 상세히 서술하라.
-        3. **기대 효과 명시**: 단순한 예측이 아니라 '매출 상승', '객단가 확보', '운영 효율화' 등 구체적인 비즈니스 이득을 강조하라.
-        4. **논리적 구조**:
-           - 1. 현재 상태 진단: 데이터 기반의 냉정한 현 상황 요약
-           - 2. 전략적 근거: 왜 이 솔루션이 지금 가장 효율적인지에 대한 분석 (`reason`과 `evaluation` 활용)
-           - 3. 최종 결과: 실행을 통해 얻게 될 정량적/정성적 기대 효과
-
-        ### 선정 및 핵심 규칙
-        - 가성비(High Impact, Low Effort)를 최우선으로 하되, 매장의 장기적인 성장 가능성을 고려하라.  
-        - 후보군이 3개 미만일 경우에만 3개보다 적게 선정하고, 그 외에는 반드시 **최적의 솔루션 3개**를 선정한다.
-        - 선정되지 않은 후보는 제외한다.
-        """),
-
-        # Few-shot: Human 예시 (SWOT + 평가 완료된 후보군)
-        ("human", f"""
-        ### 분석 데이터:
-        - SWOT 데이터: {json.dumps(example_swot, ensure_ascii=False)}
-        - 전략 후보군: {json.dumps(example_candidates, ensure_ascii=False)}
-        - 평가 결과: {json.dumps(example_evaluation, ensure_ascii=False)}
-
-        위 데이터를 바탕으로 시스템 지침(선정 기준 및 작성 규칙)에 따라 최적의 솔루션을 최대 3개 최종 선정해줘.
-        """),
-        
-        # Few-shot: AI 응답 예시 (최종 선정 결과)
-        ("ai", json.dumps(example_output, ensure_ascii=False)),
-
-        # 실제 사용자 요청
-        ("human", """
-        ### 분석 데이터:
-        - SWOT 데이터: {swot_json}
-        - 전략 후보군: {candidate_list}
-        - 평가 결과: {evaluated_candidates}
-
-        위 데이터를 바탕으로 시스템 지침(선정 기준 및 작성 규칙)에 따라 최적의 솔루션을 최대 3개 최종 선정해줘.
-        """),        
-    ])
-
-# 4. To-Do List 생성 Chain
-def get_action_plan_prompt() -> ChatPromptTemplate:
-    # 1단계: SWOT 데이터 (베이커리 예시 유지)
-    example_swot = {
-        "strengths": [{"id": "S1", "keyword": "당일 생산 수제 빵의 품질"}],
-        "weaknesses": [{"id": "W1", "keyword": "온라인 홍보 부족"}],
-        "opportunities": [{"id": "O1", "keyword": "인근 지하철역 출근길 유동인구 풍부"}],
-        "threats": [{"id": "T1", "keyword": "대형 프랜차이즈 빵집 입점 예정"}]
-    }
-
-    # 3단계에서 최종 선정된 3가지 전략 예시
-    example_selected = [
-        {
-            "id": 1,
-            "title": "출근길 타겟 '수제 모닝 샌드위치' 출시",
-            "tags": ["#매출증대", "#난이도하", "#상품개발"],
-            "related_swot": ["S1", "O1"],
-            "final_reason": "지하철역 인근의 풍부한 유동인구라는 지리적 이점을 충분히 활용하기 위해서는 출근 시간대 바쁜 직장인들을 겨냥한 특화 메뉴 도입이 시급합니다. \n매장의 핵심 경쟁력인 당일 생산 수제 빵의 품질을 샌드위치라는 간편식 형태로 전환하면, 아침 시간대 고객의 구매 허들을 낮추면서도 높은 만족도를 제공할 수 있습니다. \n이는 단순한 신규 고객 유입을 넘어 오전 시간대의 안정적인 매출 캐시카우를 확보함으로써 전체적인 수익 구조를 탄탄하게 만드는 전략적 전환점이 될 것입니다."
-        
-        },
-        {
-            "id": 2,
-            "title": "인스타그램 '빵 나오는 시간' 인증 이벤트",
-            "tags": ["#인지도제고", "#난이도하", "#마케팅"],
-            "related_swot": ["W1", "O1"],
-            "final_reason": "풍부한 유동인구에도 불구하고 온라인 노출이 부족하여 잠재 고객의 발길이 끊기는 현상을 해결하기 위해, SNS 참여형 이벤트를 통한 디지털 가시성 확보가 반드시 필요합니다. \n인스타그램을 활발히 이용하는 젊은 층을 타겟으로 빵 나오는 시간을 인증하게 유도함으로써, 별도의 마케팅 비용 없이도 자발적인 홍보 데이터를 축적하고 브랜드 인지도를 빠르게 확산시킬 수 있습니다. \n단기적인 방문 증대 효과는 물론, 온라인상의 신뢰도를 구축하여 장기적으로 매장의 경쟁력을 비약적으로 높여주는 효율적인 솔루션입니다."
-        },
-        {
-            "id": 3,
-            "title": "프랜차이즈 대비 '프리미엄 수제 빵' 브랜딩",
-            "tags": ["#신규고객", "#난이도중", "#고객관리"],
-            "related_swot": ["S1", "T1"],
-            "final_reason": "인근 대형 프랜차이즈의 입점이라는 위협 요소에 대응하기 위해, 가격 경쟁이 아닌 제품의 본질적 가치를 강조하는 프리미엄 브랜딩 전략이 핵심입니다. \n공장형 제품으로는 따라올 수 없는 당일 생산 수제 빵의 정체성을 전면에 내세움으로써 가치 소비를 중시하는 단골 고객층을 확실히 선점해야 합니다. \n브랜딩을 위한 초기 기획 과정이 수반되지만, 이는 프랜차이즈의 공세에도 흔들리지 않는 독보적인 팬층을 확보하고 장기적인 생존력을 강화하는 필수적인 투자 단계가 될 것입니다."
-        }
-    ]
-
-    # 최종 출력될 실행 계획 예시
-    example_output = [
-        {
-            "id": 1,
-            "title": "출근길 타겟 '수제 모닝 샌드위치' 출시",
-            "action_plan": [
-                {
-                    "step": 1,
-                    "title": "핵심 메뉴 구성 및 패키징 확정",
-                    "description": "기존 수제 식빵을 활용한 3분 내 서빙 가능한 레시피를 확정하고, 이동 중 취식이 편한 종이 패키지를 선정합니다.",
-                    "expected_outcome": "조리 효율성이 확보된 표준 레시피와 고객 편의성을 고려한 패키징 시안이 준비됩니다."
-                },
-                {
-                    "step": 2,
-                    "title": "얼리버드 타임 세일 프로모션 세팅",
-                    "description": "오전 7시~9시 사이 샌드위치 구매 시 아메리카노 할인 등 출근족을 위한 가격 혜택을 설계합니다.",
-                    "expected_outcome": "경쟁사 대비 아침 시간대 방문 유인책을 확보하여 신규 유입을 유도합니다."
-                },
-                {
-                    "step": 3,
-                    "title": "매장 외부 가시성 강화(X-배너 설치)",
-                    "description": "지하철역 출구 방향 유동인구가 볼 수 있도록 매장 입구에 '갓 구운 모닝 샌드위치' 홍보 배너를 설치합니다.",
-                    "expected_outcome": "매장 앞 유입률과 아침 시간대 신규 고객 수가 즉각적으로 증가합니다."
-                }
-            ]
-        },
-        {
-            "id": 2,
-            "title": "인스타그램 '빵 나오는 시간' 인증 이벤트",
-            "action_plan": [
-                {
-                    "step": 1,
-                    "title": "이벤트 홍보물 및 참여 가이드 제작",
-                    "description": "매장 내 '빵 나오는 시간표'와 함께 SNS 인증 시 혜택(예: 쿠키 증정)을 알리는 안내문을 비치합니다.",
-                    "expected_outcome": "내방 고객들이 이벤트 내용을 인지하고 참여할 수 있는 환경이 조성됩니다."
-                },
-                {
-                    "step": 2,
-                    "title": "공식 계정 활성화 및 해시태그 설정",
-                    "description": "지역 맛집 관련 해시태그를 설정하고, 고객이 올린 인증샷을 공식 계정에서 리포스팅하며 소통합니다.",
-                    "expected_outcome": "지역 기반 검색 결과에 매장 노출 빈도가 높아지며 온라인 바이럴이 시작됩니다."
-                },
-                {
-                    "step": 3,
-                    "title": "베스트 리뷰어 선정 및 재방문권 증정",
-                    "description": "매주 정성스러운 후기를 남긴 고객 3명을 선정하여 재방문 시 사용 가능한 할인권을 DM으로 발송합니다.",
-                    "expected_outcome": "고품질 후기 데이터가 축적되며 고객의 자발적인 재방문율이 상승합니다."
-                }
-            ]
-        },
-        {
-            "id": 3,
-            "title": "프랜차이즈 대비 '프리미엄 수제 빵' 브랜딩",
-            "action_plan": [
-                {
-                    "step": 1,
-                    "title": "장인 정신 소구 문구 및 스토리 개발",
-                    "description": "'24시간 저온 숙성', '천연 효모 사용' 등 제품의 차별화된 가치를 담은 브랜드 카피를 개발합니다.",
-                    "expected_outcome": "프랜차이즈와 다른 우리 매장만의 고유한 브랜드 메시지가 정립됩니다."
-                },
-                {
-                    "step": 2,
-                    "title": "프리미엄 패키징 및 진열 구획 개선",
-                    "description": "시그니처 제품에 고급스러운 라벨지를 부착하고, 조명과 진열대를 조정하여 제품이 돋보이게 연출합니다.",
-                    "expected_outcome": "고객이 체감하는 시각적 품질이 향상되어 가치 소비를 지향하는 고객을 만족시킵니다."
-                },
-                {
-                    "step": 3,
-                    "title": "멤버십 포인트 및 단골 우대 혜택 강화",
-                    "description": "단골 고객을 위한 선결제 시스템이나 신메뉴 우선 시식 기회를 제공하는 VIP 등급제를 도입합니다.",
-                    "expected_outcome": "프랜차이즈 침투에도 이탈하지 않는 강력한 로열티를 가진 단골 팬층이 형성됩니다."
-                }
-            ]
-        }
-    ]
-
-    # AI 프롬프트
-    return ChatPromptTemplate.from_messages([
-        ("system", """
-        너는 선정된 비즈니스 전략을 실질적인 운영 프로세스로 전환하는 '현장 중심형 실행 컨설턴트'다. 
-        사장님이 복잡한 고민 없이 즉시 실행할 수 있도록 3단계의 구체적인 액션 아이템을 도출하라.
-
-        ### 목표
-        최종 선정된 각 전략에 대해 **[준비 - 실행 - 고도화]**의 흐름을 가진 3단계 실행 계획(Action Plan)을 수립하라.
-
-        ### 작성 규칙
-        1. **현실 가능성**: 소상공인의 한정된 인력과 예산으로도 일주일 내에 착수할 수 있는 현실적인 방안이어야 한다.
-        2. **구체성**: '홍보를 잘 하자' 같은 추상적인 문구는 배제하고, '어떤 매체를 통해', '어떤 문구로', '어떻게' 할지 명확히 서술하라.
-        3. **단계별 흐름**:
-            - 1단계(준비): 상품 기획, 재료 준비, 인프라 세팅 등 기초 단계
-            - 2단계(실행): 실제 출시, 프로모션 진행, 고객 응대 등 현장 적용 단계
-            - 3단계(확산/관리): 후기 유도, 단골 관리, 데이터 분석 등 지속 가능 단계
-        4. **가시적 결과 명시**: 각 단계(`step`)마다 해당 행동을 완료했을 때 사장님이 얻을 수 있는 '기대 결과(`expected_outcome`)'를 한 문장으로 구체적으로 포함하라.
-        5. **어조**: 신뢰감을 주는 비즈니스 어투를 사용하되, 실천을 독려하는 긍정적인 뉘앙스를 유지하라.
-        """),
-
-        # Few-shot 데이터
-        ("human", f"""
-        ### 분석 데이터:
-        - SWOT 데이터: {json.dumps(example_swot, ensure_ascii=False)}
-        - 최종 선정 전략: {json.dumps(example_selected, ensure_ascii=False)}
-
-        위 전략들에 대해 사장님이 바로 행동에 옮길 수 있는 3단계 실행 계획을 세밀하게 작성해줘.
-        """),
-        
-        ("ai", json.dumps(example_output, ensure_ascii=False)),
-
-        # 실제 요청
-        ("human", """
-        ### 분석 데이터:
-        - SWOT 데이터: {swot_json}
-        - 최종 선정 전략: {final_selected_strategies}
-
-        위 전략들에 대해 사장님이 바로 행동에 옮길 수 있는 3단계 실행 계획을 세밀하게 작성해줘.
-        """),
-    ])
