+++ app/core/prompt_templates/candidate_prompt.py
@@ -1,0 +1,87 @@
+import json
+from langchain_core.prompts import ChatPromptTemplate
+
+# 1. 후보 전략 선정 Chain
+def get_candidate_prompt() -> ChatPromptTemplate:
+    # Few-shot Input (베이커리 예시)
+    example_input = {
+        "strengths": [{"id": "S1", "keyword": "당일 생산 수제 빵의 품질"}],
+        "weaknesses": [{"id": "W1", "keyword": "온라인 홍보 부족"}],
+        "opportunities": [{"id": "O1", "keyword": "인근 지하철역 출근길 유동인구 풍부"}],
+        "threats": [{"id": "T1", "keyword": "대형 프랜차이즈 빵집 입점 예정"}]
+    }
+    # Few-shot Output
+    example_output = [
+        {
+            "id": 1,
+            "title": "출근길 타겟 '수제 모닝 샌드위치' 출시",
+            "tags": ["#매출증대", "#난이도하", "#상품개발"],
+            "related_swot": ["S1", "O1"],
+            "reason": "수제 빵의 품질(S1)과 풍부한 출근길 유동인구(O1)를 결합하여 아침 시간대 구매를 유도하는 SO 전략이다."
+        },
+        {
+            "id": 2,
+            "title": "인스타그램 '빵 나오는 시간' 인증 이벤트",
+            "tags": ["#인지도제고", "#난이도하", "#마케팅"],
+            "related_swot": ["W1", "O1"],
+            "reason": "홍보 부족(W1)을 해결하기 위해 유동인구(O1)를 대상으로 SNS 참여를 유도하여 온라인 노출을 극대화하는 WO 전략이다."
+        },
+        {
+            "id": 3,
+            "title": "프랜차이즈 대비 '프리미엄 수제 빵' 브랜딩",
+            "tags": ["#신규고객", "#난이도중", "#고객관리"],
+            "related_swot": ["S1", "T1"],
+            "reason": "대형 프랜차이즈(T1)에 맞서 수제 빵의 품질(S1)을 강조하여 장인 정신이 담긴 매장임을 차별화하는 ST 전략이다."
+        },
+        {
+            "id": 4,
+            "title": "인근 지역 커뮤니티 타겟 마케팅 강화",
+            "tags": ["#바이럴", "#난이도중", "#마케팅"],
+            "related_swot": ["W1", "T1"],
+            "reason": "부족한 홍보(W1)를 보완하고 경쟁사(T1)의 유입 독점을 막기 위해 지역 맘카페 등에 신뢰 기반 마케팅을 펼치는 WT 전략이다."
+        },
+        {
+            "id": 5,
+            "title": "출근 시간대 '사전 예약 및 픽업' 서비스",
+            "tags": ["#운영효율", "#난이도중", "#시설개선"],
+            "related_swot": ["S1", "O1"],
+            "reason": "빵의 품질(S1)을 아는 바쁜 직장인(O1)을 위해 대기 시간을 줄여 편의성을 제공하는 고도화된 SO 전략이다."
+        }
+    ]
+    
+    # AI 프롬프트
+    return ChatPromptTemplate.from_messages([
+        ("system", """
+        너는 주어진 SWOT 분석 결과를 바탕으로, 소상공인을 위한 구체적이고 실행 가능한 비즈니스 전략을 제안하는 최고의 데이터 기반 전략 컨설턴트다.
+
+        ### 목표
+        입력된 SWOT 분석(JSON)을 바탕으로, 사용자가 실행할 수 있는 후보 솔루션을 3~5개 생성하라. 
+        이 단계에서는 최종 선정 전, 다양한 전략적 조합(SO, WO, ST, WT)을 통해 아이디어를 발산하는 것에 집중한다.
+
+        ### 핵심 규칙
+        1. **전략적 사고**: SO(강점-기회), WO(약점-기회), ST(강점-위협), WT(약점-위협) 등의 조합을 활용하여 솔루션을 도출하라.
+        2. **구체성**: `title`은 사용자가 즉시 이해할 수 있도록 구체적인 액션을 담아 30자 이내로 작성하라.
+        3. **구조화된 태그**: `tags`는 '#목표', '#난이도', '#카테고리' 순서로 반드시 3개를 포함하라.
+            - #목표 예시: #매출증대, #신규고객, #객단가UP, #인지도제고
+            - #난이도 예시: #난이도하, #난이도중, #난이도상
+            - #카테고리 예시: #마케팅, #운영, #고객관리, #상품개발, #시설개선
+        4. **근거 제시**: `related_swot`에는 솔루션 도출에 가장 큰 영향을 준 SWOT 항목의 `id`를 배열로 포함하라.
+        5. **생성 이유 명시**: `reason` 필드에 해당 SWOT 항목들을 왜 그렇게 조합했는지, 그리고 이 전략이 어떤 기대 효과를 갖는지 2문장 내외로 친절하게 설명하라.
+        6. **ID 생성**: Id는 1부터 시작하여 생성되는 후보군의 순서에 따라 **1씩 증가하는 정수(Integer)**로 부여하라. (예: 1, 2, 3...)
+        7. **후보군 개수**: 반드시 3개 이상, 5개 이하의 솔루션을 생성하라.
+        """),
+
+        # Few-shot: Human 예시
+        ("human", f"### 입력된 SWOT 분석 데이터:\n{json.dumps(example_input, ensure_ascii=False)}"),
+        
+        # Few-shot: AI 응답 예시
+        ("ai", json.dumps(example_output, ensure_ascii=False)),
+
+        # 실제 요청
+        ("human", """
+        ### 입력된 SWOT 분석 데이터:
+        {swot_json}        
+
+        이제 위 형식에 맞춰 후보군을 생성해줘.
+        """),
+    ])
