+++ app/core/prompt_templates/final_select_prompt.py
@@ -1,0 +1,141 @@
+import json
+from langchain_core.prompts import ChatPromptTemplate
+
+# 3. 최종 선정 Chain
+def get_selection_prompt() -> ChatPromptTemplate:
+    # 1단계: SWOT 데이터 정의 (베이커리 예시)
+    example_swot = {
+        "strengths": [{"id": "S1", "keyword": "당일 생산 수제 빵의 품질"}],
+        "weaknesses": [{"id": "W1", "keyword": "온라인 홍보 부족"}],
+        "opportunities": [{"id": "O1", "keyword": "인근 지하철역 출근길 유동인구 풍부"}],
+        "threats": [{"id": "T1", "keyword": "대형 프랜차이즈 빵집 입점 예정"}]
+    }
+
+    # 2단계: 1번 프롬프트에서 생성된 후보군 (5개)
+    example_candidates = [
+        {
+            "id": 1,
+            "title": "출근길 타겟 '수제 모닝 샌드위치' 출시",
+            "tags": ["#매출증대", "#난이도하", "#상품개발"],
+            "related_swot": ["S1", "O1"],
+            "reason": "수제 빵의 품질(S1)과 풍부한 출근길 유동인구(O1)를 결합하여 아침 시간대 구매를 유도하는 SO 전략이다."
+        },
+        {
+            "id": 2,
+            "title": "인스타그램 '빵 나오는 시간' 인증 이벤트",
+            "tags": ["#인지도제고", "#난이도하", "#마케팅"],
+            "related_swot": ["W1", "O1"],
+            "reason": "홍보 부족(W1)을 해결하기 위해 유동인구(O1)를 대상으로 SNS 참여를 유도하여 온라인 노출을 극대화하는 WO 전략이다."
+        },
+        {
+            "id": 3,
+            "title": "프랜차이즈 대비 '프리미엄 수제 빵' 브랜딩",
+            "tags": ["#신규고객", "#난이도중", "#고객관리"],
+            "related_swot": ["S1", "T1"],
+            "reason": "대형 프랜차이즈(T1)에 맞서 수제 빵의 품질(S1)을 강조하여 장인 정신이 담긴 매장임을 차별화하는 ST 전략이다."
+        },
+        {
+            "id": 4,
+            "title": "인근 지역 커뮤니티 타겟 마케팅 강화",
+            "tags": ["#바이럴", "#난이도중", "#마케팅"],
+            "related_swot": ["W1", "T1"],
+            "reason": "부족한 홍보(W1)를 보완하고 경쟁사(T1)의 유입 독점을 막기 위해 지역 맘카페 등에 신뢰 기반 마케팅을 펼치는 WT 전략이다."
+        },
+        {
+            "id": 5,
+            "title": "출근 시간대 '사전 예약 및 픽업' 서비스",
+            "tags": ["#운영효율", "#난이도중", "#시설개선"],
+            "related_swot": ["S1", "O1"],
+            "reason": "빵의 품질(S1)을 아는 바쁜 직장인(O1)을 위해 대기 시간을 줄여 편의성을 제공하는 고도화된 SO 전략이다."
+        }
+    ]
+
+    # 3단계: 2번 프롬프트에서 생성된 평가 결과 (5개)
+    example_evaluation = [
+        {"id": 1, "impactScore": 9, "effortScore": 2, "evaluation": "이미 확보된 품질과 유동인구를 즉시 매출로 연결하며, 메뉴 구성 외 추가 비용이 적어 효율적이다."},
+        {"id": 2, "impactScore": 7, "effortScore": 1, "evaluation": "마케팅 비용 없이 고객 참여만으로 온라인 인지도를 빠르게 쌓을 수 있어 실행이 매우 쉽다."},
+        {"id": 3, "impactScore": 8, "effortScore": 4, "evaluation": "경쟁사와 차별화되는 장기적 자산이 되지만, 브랜딩을 위한 스토리텔링과 디자인 작업이 수반되어야 한다."},
+        {"id": 4, "impactScore": 6, "effortScore": 3, "evaluation": "지역 밀착형 홍보로 확실한 단골을 잡을 수 있으나, 커뮤니티 관리와 신뢰 형성에 시간이 소요된다."},
+        {"id": 5, "impactScore": 5, "effortScore": 6, "evaluation": "편의성은 증대되나 예약 관리 시스템 도입과 인력 동선 조정 등 운영상 복잡성이 발생할 수 있다."}
+    ]
+
+    # 최종 결과물: 가성비(Impact 높고 Effort 낮은) 및 다양성 고려한 상위 3개 선정
+    example_output = [
+        {
+            "id": 1,
+            "title": "출근길 타겟 '수제 모닝 샌드위치' 출시",
+            "tags": ["#매출증대", "#난이도하", "#상품개발"],
+            "related_swot": ["S1", "O1"],
+            "final_reason": "지하철역 인근의 풍부한 유동인구라는 지리적 이점을 충분히 활용하기 위해서는 출근 시간대 바쁜 직장인들을 겨냥한 특화 메뉴 도입이 시급합니다. \n매장의 핵심 경쟁력인 당일 생산 수제 빵의 품질을 샌드위치라는 간편식 형태로 전환하면, 아침 시간대 고객의 구매 허들을 낮추면서도 높은 만족도를 제공할 수 있습니다. \n이는 단순한 신규 고객 유입을 넘어 오전 시간대의 안정적인 매출 캐시카우를 확보함으로써 전체적인 수익 구조를 탄탄하게 만드는 전략적 전환점이 될 것입니다."
+        
+        },
+        {
+            "id": 2,
+            "title": "인스타그램 '빵 나오는 시간' 인증 이벤트",
+            "tags": ["#인지도제고", "#난이도하", "#마케팅"],
+            "related_swot": ["W1", "O1"],
+            "final_reason": "풍부한 유동인구에도 불구하고 온라인 노출이 부족하여 잠재 고객의 발길이 끊기는 현상을 해결하기 위해, SNS 참여형 이벤트를 통한 디지털 가시성 확보가 반드시 필요합니다. \n인스타그램을 활발히 이용하는 젊은 층을 타겟으로 빵 나오는 시간을 인증하게 유도함으로써, 별도의 마케팅 비용 없이도 자발적인 홍보 데이터를 축적하고 브랜드 인지도를 빠르게 확산시킬 수 있습니다. \n단기적인 방문 증대 효과는 물론, 온라인상의 신뢰도를 구축하여 장기적으로 매장의 경쟁력을 비약적으로 높여주는 효율적인 솔루션입니다."
+        },
+        {
+            "id": 3,
+            "title": "프랜차이즈 대비 '프리미엄 수제 빵' 브랜딩",
+            "tags": ["#신규고객", "#난이도중", "#고객관리"],
+            "related_swot": ["S1", "T1"],
+            "final_reason": "인근 대형 프랜차이즈의 입점이라는 위협 요소에 대응하기 위해, 가격 경쟁이 아닌 제품의 본질적 가치를 강조하는 프리미엄 브랜딩 전략이 핵심입니다. \n공장형 제품으로는 따라올 수 없는 당일 생산 수제 빵의 정체성을 전면에 내세움으로써 가치 소비를 중시하는 단골 고객층을 확실히 선점해야 합니다. \n브랜딩을 위한 초기 기획 과정이 수반되지만, 이는 프랜차이즈의 공세에도 흔들리지 않는 독보적인 팬층을 확보하고 장기적인 생존력을 강화하는 필수적인 투자 단계가 될 것입니다."
+        }
+    ]
+
+    
+    # AI 프롬프트
+    return ChatPromptTemplate.from_messages([
+        ("system", """
+        너는 여러 비즈니스 전략 중 소상공인에게 가장 큰 이익을 줄 수 있는 핵심 솔루션을 최대 3개까지 선정하는 수석 전략 컨설턴트다.
+
+        ### 목표
+        평가된 후보군 중 가성비(High Impact, Low Effort)가 뛰어난 **핵심 솔루션(Core Solutions)**을 우선순위에 따라 **최대 3개** 선정하여 리스트 형식으로 반환하라.
+
+        ### 선정 기준
+        1. **제 1원칙 (Impact)**: 기대 효과(Impact Score)가 가장 높은 솔루션을 우선한다.
+        2. **제 2원칙 (Effort)**: 기대 효과 점수가 동일하다면, 실행 난이도(Effort Score)가 가장 낮은 솔루션을 선택한다.
+        3. **제 3원칙 (Diversity)**: 가능한 한 서로 다른 SWOT 요소(S, W, O, T)를 해결하는 솔루션들을 조합하여 다각적인 전략을 구성하라.
+        3. **최종 검증**: 선정된 솔루션이 주어진 SWOT 데이터와 논리적으로 완벽하게 연결되는지 확인하라.
+         
+        ### final_reason 작성 지침 (분석적 접근)
+        사장님이 전략의 타당성을 이성적으로 납득할 수 있도록 다음 규칙을 엄수하라.
+        1. **전문적이고 정중한 톤**: 예의는 갖추되, 감성적인 위로보다는 전문가적인 통찰을 전달하라.
+        2. **논리적 인과관계**: '강점 활용 - 기회 포착 - 약점 보완' 등 SWOT 분석 결과가 어떻게 수익으로 연결되는지 4~5줄로 상세히 서술하라.
+        3. **기대 효과 명시**: 단순한 예측이 아니라 '매출 상승', '객단가 확보', '운영 효율화' 등 구체적인 비즈니스 이득을 강조하라.
+        4. **논리적 구조**:
+           - 1. 현재 상태 진단: 데이터 기반의 냉정한 현 상황 요약
+           - 2. 전략적 근거: 왜 이 솔루션이 지금 가장 효율적인지에 대한 분석 (`reason`과 `evaluation` 활용)
+           - 3. 최종 결과: 실행을 통해 얻게 될 정량적/정성적 기대 효과
+
+        ### 선정 및 핵심 규칙
+        - 가성비(High Impact, Low Effort)를 최우선으로 하되, 매장의 장기적인 성장 가능성을 고려하라.  
+        - 후보군이 3개 미만일 경우에만 3개보다 적게 선정하고, 그 외에는 반드시 **최적의 솔루션 3개**를 선정한다.
+        - 선정되지 않은 후보는 제외한다.
+        """),
+
+        # Few-shot: Human 예시 (SWOT + 평가 완료된 후보군)
+        ("human", f"""
+        ### 분석 데이터:
+        - SWOT 데이터: {json.dumps(example_swot, ensure_ascii=False)}
+        - 전략 후보군: {json.dumps(example_candidates, ensure_ascii=False)}
+        - 평가 결과: {json.dumps(example_evaluation, ensure_ascii=False)}
+
+        위 데이터를 바탕으로 시스템 지침(선정 기준 및 작성 규칙)에 따라 최적의 솔루션을 최대 3개 최종 선정해줘.
+        """),
+        
+        # Few-shot: AI 응답 예시 (최종 선정 결과)
+        ("ai", json.dumps(example_output, ensure_ascii=False)),
+
+        # 실제 사용자 요청
+        ("human", """
+        ### 분석 데이터:
+        - SWOT 데이터: {swot_json}
+        - 전략 후보군: {candidate_list}
+        - 평가 결과: {evaluated_candidates}
+
+        위 데이터를 바탕으로 시스템 지침(선정 기준 및 작성 규칙)에 따라 최적의 솔루션을 최대 3개 최종 선정해줘.
+        """),        
+    ])
